pyproject.toml
[project]
name = "smart-log-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
